Refactor started, added test for slack markdown infrastructure
